Convert to minitest plugin.
This allows minitest-emoji to work with minitest 5.
