+++ Manifest.txt
@@ -4,4 +4,5 @@ Manifest.txt
 README.rdoc
 Rakefile
 lib/minitest/emoji.rb
+lib/minitest/emoji_plugin.rb
 test/test_minitest_emoji.rb
